WIP: Bump mdsal to 5.0.0
This bumps mdsal to 5.0.0 as part of Magnesium MRI.

+++ openflowplugin/src/test/java/org/opendaylight/openflowplugin/openflow/md/core/sal/convertor/match/MatchConvertorV10Test.java
@@ -41,6 +41,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026
 import org.opendaylight.yang.gen.v1.urn.opendaylight.model.match.types.rev131026.vlan.match.fields.VlanIdBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.FlowWildcardsV10;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.match.v10.grouping.MatchV10;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Created by Martin Bobak mbobak@cisco.com on 8/30/14.
@@ -155,7 +156,7 @@ public class MatchConvertorV10Test {
 
         // Specify ICMP code only.
         Integer icmpCode = 31;
-        icmpv4MatchBuilder = new Icmpv4MatchBuilder().setIcmpv4Type(null).setIcmpv4Code(icmpCode.shortValue());
+        icmpv4MatchBuilder = new Icmpv4MatchBuilder().setIcmpv4Type((Uint8) null).setIcmpv4Code(icmpCode.shortValue());
         wcTpSrc = true;
         wcTpDst = false;
         wc = new FlowWildcardsV10(
